pyproject.toml
[project]
name = "basecamp-ai-sec"
version = "0.1.0"
description = "Structural tests for the Basecamp AI + Cybersecurity learning map"
requires-python = ">=3.11"
dependencies = [
    "pyyaml>=6.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
]

[tool.setuptools]
# Prevent setuptools from auto-discovering topics/, paths/, and scripts/
# as Python packages. This is a content repo, not a library — the only
# Python code lives in tests/ and is invoked directly by pytest.
packages = []

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
addopts = "-ra --strict-markers"
